Refactor to accept pods vs. Lister
